+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v0.2.20
-- New feature: infinite mining stamina
-- Changed infinite farming/building/mining stamina to only restore the necessary stamina for the action
+- New feature: infinite mining/wood cutting stamina
+- Changed infinite farming/building/mining/wood cutting stamina to only restore the necessary stamina for the action
+- Potential fix for Mod sometimes stops working after a player disconnects [#45](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/45)
 
 ### v0.2.19
 - New features: